+++ admin/views/dashboard/tmpl/default.php
@@ -3,7 +3,7 @@
  * @package      CrowdFunding
  * @subpackage   Components
  * @author       Todor Iliev
- * @copyright    Copyright (C) 2014 Todor Iliev <todor@itprism.com>. All rights reserved.
+ * @copyright    Copyright (C) 2015 Todor Iliev <todor@itprism.com>. All rights reserved.
  * @license      http://www.gnu.org/copyleft/gpl.html GNU/GPL
  */
 
@@ -151,8 +151,14 @@ defined('_JEXEC') or die;
         </div>
 
         <div class="span4">
-            <a href="http://itprism.com/free-joomla-extensions/ecommerce-gamification/crowdfunding-collective-raising-capital" target="_blank"><img src="../media/com_crowdfunding/images/logo.png" alt="<?php echo JText::_("COM_CROWDFUNDING"); ?>"/></a>
-            <a href="http://itprism.com" target="_blank" title="<?php echo JText::_("COM_CROWDFUNDING_PRODUCT"); ?>"><img src="../media/com_crowdfunding/images/product_of_itprism.png" alt="<?php echo JText::_("COM_CROWDFUNDING_PRODUCT"); ?>"/></a>
+            <a href="http://itprism.com/free-joomla-extensions/ecommerce-gamification/crowdfunding-collective-raising-capital"
+               target="_blank"><img src="../media/com_crowdfunding/images/logo.png"
+                                    alt="<?php echo JText::_("COM_CROWDFUNDING"); ?>"/></a>
+            <a href="http://itprism.com" target="_blank"
+               title="<?php echo JText::_("COM_CROWDFUNDING_PRODUCT"); ?>"><img
+                    src="../media/com_crowdfunding/images/product_of_itprism.png"
+                    alt="<?php echo JText::_("COM_CROWDFUNDING_PRODUCT"); ?>"/></a>
+
             <p><?php echo JText::_("COM_CROWDFUNDING_YOUR_VOTE"); ?></p>
             <p><?php echo JText::_("COM_CROWDFUNDING_SPONSORSHIP"); ?></p>
             <p><?php echo JText::_("COM_CROWDFUNDING_SUBSCRIPTION"); ?></p>